+++ src/resources/v1/events/credits/credits.ts
@@ -1,6 +1,20 @@
 // File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
 import { APIResource } from '../../../../core/resource';
+import * as CustomCurrenciesAPI from './custom-currencies';
+import {
+  CustomCurrencies,
+  CustomCurrencyArchiveResponse,
+  CustomCurrencyCreateParams,
+  CustomCurrencyCreateResponse,
+  CustomCurrencyListAssociatedEntitiesResponse,
+  CustomCurrencyListParams,
+  CustomCurrencyListResponse,
+  CustomCurrencyListResponsesMyCursorIDPage,
+  CustomCurrencyUnarchiveResponse,
+  CustomCurrencyUpdateParams,
+  CustomCurrencyUpdateResponse,
+} from './custom-currencies';
 import * as GrantsAPI from './grants';
 import {
   CreditGrantResponse,
@@ -16,6 +30,9 @@ import { RequestOptions } from '../../../../internal/request-options';
 
 export class Credits extends APIResource {
   grants: GrantsAPI.Grants = new GrantsAPI.Grants(this._client);
+  customCurrencies: CustomCurrenciesAPI.CustomCurrencies = new CustomCurrenciesAPI.CustomCurrencies(
+    this._client,
+  );
 
   /**
    * Retrieves the automatic recharge configuration for a customer and currency.
@@ -347,6 +364,7 @@ export interface CreditListLedgerParams extends MyCursorIDPageParams {
 }
 
 Credits.Grants = Grants;
+Credits.CustomCurrencies = CustomCurrencies;
 
 export declare namespace Credits {
   export {
@@ -366,5 +384,19 @@ export declare namespace Credits {
     type GrantListResponsesMyCursorIDPage as GrantListResponsesMyCursorIDPage,
     type GrantCreateParams as GrantCreateParams,
     type GrantListParams as GrantListParams,
+  };
+
+  export {
+    CustomCurrencies as CustomCurrencies,
+    type CustomCurrencyCreateResponse as CustomCurrencyCreateResponse,
+    type CustomCurrencyUpdateResponse as CustomCurrencyUpdateResponse,
+    type CustomCurrencyListResponse as CustomCurrencyListResponse,
+    type CustomCurrencyArchiveResponse as CustomCurrencyArchiveResponse,
+    type CustomCurrencyListAssociatedEntitiesResponse as CustomCurrencyListAssociatedEntitiesResponse,
+    type CustomCurrencyUnarchiveResponse as CustomCurrencyUnarchiveResponse,
+    type CustomCurrencyListResponsesMyCursorIDPage as CustomCurrencyListResponsesMyCursorIDPage,
+    type CustomCurrencyCreateParams as CustomCurrencyCreateParams,
+    type CustomCurrencyUpdateParams as CustomCurrencyUpdateParams,
+    type CustomCurrencyListParams as CustomCurrencyListParams,
   };
 }