+++ src/resources/v1/events/credits/index.ts
@@ -11,6 +11,19 @@ export {
   type CreditListLedgerResponsesMyCursorIDPage,
 } from './credits';
 export {
+  CustomCurrencies,
+  type CustomCurrencyCreateResponse,
+  type CustomCurrencyUpdateResponse,
+  type CustomCurrencyListResponse,
+  type CustomCurrencyArchiveResponse,
+  type CustomCurrencyListAssociatedEntitiesResponse,
+  type CustomCurrencyUnarchiveResponse,
+  type CustomCurrencyCreateParams,
+  type CustomCurrencyUpdateParams,
+  type CustomCurrencyListParams,
+  type CustomCurrencyListResponsesMyCursorIDPage,
+} from './custom-currencies';
+export {
   Grants,
   type CreditGrantResponse,
   type GrantListResponse,