+++ tests/api-resources/v1/addons/addons.test.ts
@@ -91,7 +91,7 @@ describe('resource addons', () => {
           },
           limit: 1,
           productId: 'productId',
-          status: 'status',
+          status: ['DRAFT'],
         },
         { path: '/_stainless_unknown_path' },
       ),