+++ tests/api-resources/v1/customers/promotional-entitlements.test.ts
@@ -87,7 +87,7 @@ describe('resource promotionalEntitlements', () => {
             lte: '2019-12-27T18:11:19.117Z',
           },
           limit: 1,
-          status: 'status',
+          status: ['Active'],
         },
         { path: '/_stainless_unknown_path' },
       ),